requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==1.1.0
markdown==3.5.1
openpyxl==3.1.2
xlrd==2.0.1
